feat(production-statuses): per-product stages ui with consumption editor
