+++ resources/js/pages/production-statuses/components/add-status-form.tsx
