+++ resources/js/pages/stock/form.ts
@@ -1,5 +1,4 @@
 export type FormData = Pick<Stockable, 'name' | 'unit'> & {
-    products: number[];
     alert_at: string;
     quantity: string;
 };
