+++ resources/js/pages/tracking/index.tsx
@@ -1,64 +1,21 @@
-import { Badge } from '@/components/ui/badge';
-import { Button } from '@/components/ui/button';
 import {
     Card,
     CardDescription,
     CardHeader,
     CardTitle,
 } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import {
-    Select,
-    SelectContent,
-    SelectItem,
-    SelectTrigger,
-    SelectValue,
-} from '@/components/ui/select';
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow,
-} from '@/components/ui/table';
 import AppLayout from '@/layouts/app-layout';
-import { capitalize } from '@/lib/utils';
 import { BreadcrumbItem } from '@/types';
 import { Head, router } from '@inertiajs/react';
-import { ArrowRight, Flame, Search } from 'lucide-react';
-import { useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { toast } from 'sonner';
-import { nextStatusFor, toggleGroup, toggleId } from './selection';
-
-type TrackingDetail = {
-    id: number;
-    order_id: number;
-    child_name: string | null;
-    client_name: string;
-    school: string;
-    classroom: string;
-    photo_number: number | null;
-    attended_photo_session: boolean | null;
-    product_id: number;
-    product_name: string;
-    product_type_id: number;
-    product_type: string | null;
-    variant: Record<string, string> | null;
-    note: string | null;
-    production_status_id: number | null;
-    production_status: string | null;
-    position: number;
-    priority: boolean;
-    status_updated_at: string | null;
-};
-
-type Filters = {
-    search: string | null;
-    school_id: string | null;
-    product_type_id: string | null;
-    production_status_id: string | null;
-};
+import {
+    ProductGroup,
+    TrackingDetail,
+    TrackingProduct,
+} from './components/product-group';
+import { Filters, TrackingFilters } from './components/tracking-filters';
+import { useSelection } from './hooks/use-selection';
 
 const breadcrumbs: BreadcrumbItem[] = [
     {
@@ -69,32 +26,35 @@ const breadcrumbs: BreadcrumbItem[] = [
 
 export default function Tracking({
     details,
+    products,
     productTypes,
     schools,
     filters,
 }: PageProps<{
     details: TrackingDetail[];
-    productTypes: ProductTypeWithStatuses[];
+    products: TrackingProduct[];
+    productTypes: ProductType[];
     schools: Array<{ id: number; name: string }>;
     filters: Filters;
 }>) {
-    const [selected, setSelected] = useState<number[]>([]);
-    const [search, setSearch] = useState(filters.search ?? '');
+    const { selected, toggle, toggleGroupItems, clear } = useSelection();
 
     const groups = useMemo(() => {
-        return productTypes
-            .map((type) => ({
-                type,
+        return products
+            .map((product) => ({
+                product,
                 items: details.filter(
-                    (detail) => detail.product_type_id === type.id,
+                    (detail) => detail.product_id === product.id,
                 ),
             }))
             .filter((group) => group.items.length > 0);
-    }, [details, productTypes]);
+    }, [details, products]);
 
-    const applyFilters = (overrides: Partial<Filters>) => {
+    const applyFilters = (
+        overrides: Partial<Filters> & { search?: string },
+    ) => {
         const params: Record<string, string> = {};
-        const next = { ...filters, search, ...overrides };
+        const next = { ...filters, ...overrides };
 
         if (next.search) params.search = next.search;
         if (next.school_id) params.school_id = String(next.school_id);
@@ -107,16 +67,6 @@ export default function Tracking({
             preserveState: true,
             preserveScroll: true,
         });
-    };
-
-    const toggle = (id: number) => {
-        setSelected((prev) => toggleId(prev, id));
-    };
-
-    const toggleGroupItems = (items: TrackingDetail[]) => {
-        const ids = items.map((item) => item.id);
-
-        setSelected((prev) => toggleGroup(prev, ids));
     };
 
     const applyStatus = (
@@ -138,7 +88,7 @@ export default function Tracking({
             {
                 preserveScroll: true,
                 onSuccess: () => {
-                    setSelected([]);
+                    clear();
                     toast.success(
                         statusName
                             ? `Estado actualizado a "${statusName}"`
@@ -160,83 +110,12 @@ export default function Tracking({
             <Head title="Seguimiento" />
 
             <section className="flex flex-col gap-6 p-6">
-                <div className="flex flex-col gap-2 md:flex-row md:items-center">
-                    <form
-                        className="flex items-center gap-2"
-                        onSubmit={(e) => {
-                            e.preventDefault();
-                            applyFilters({ search });
-                        }}
-                    >
-                        <Input
-                            placeholder="Niño, cliente o n° de pedido"
-                            value={search}
-                            onChange={(e) => setSearch(e.target.value)}
-                            className="w-64"
-                        />
-                        <Button type="submit" variant="secondary" size="icon">
-                            <Search className="h-4 w-4" />
-                        </Button>
-                    </form>
-
-                    <Select
-                        value={
-                            filters.school_id
-                                ? String(filters.school_id)
-                                : 'all'
-                        }
-                        onValueChange={(value) =>
-                            applyFilters({
-                                school_id: value === 'all' ? null : value,
-                            })
-                        }
-                    >
-                        <SelectTrigger className="md:w-56">
-                            <SelectValue placeholder="Escuela" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            <SelectItem value="all">
-                                Todas las escuelas
-                            </SelectItem>
-                            {schools.map((school) => (
-                                <SelectItem
-                                    value={String(school.id)}
-                                    key={school.id}
-                                >
-                                    {school.name}
-                                </SelectItem>
-                            ))}
-                        </SelectContent>
-                    </Select>
-
-                    <Select
-                        value={
-                            filters.product_type_id
-                                ? String(filters.product_type_id)
-                                : 'all'
-                        }
-                        onValueChange={(value) =>
-                            applyFilters({
-                                product_type_id: value === 'all' ? null : value,
-                            })
-                        }
-                    >
-                        <SelectTrigger className="md:w-48">
-                            <SelectValue placeholder="Tipo de producto" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            <SelectItem value="all">Todos los tipos</SelectItem>
-                            {productTypes.map((type) => (
-                                <SelectItem
-                                    value={String(type.id)}
-                                    key={type.id}
-                                >
-                                    {capitalize(type.name)}
-                                </SelectItem>
-                            ))}
-                        </SelectContent>
-                    </Select>
-                </div>
+                <TrackingFilters
+                    filters={filters}
+                    schools={schools}
+                    productTypes={productTypes}
+                    onApply={applyFilters}
+                />
 
                 {groups.length === 0 ? (
                     <Card>
@@ -249,244 +128,21 @@ export default function Tracking({
                         </CardHeader>
                     </Card>
                 ) : (
-                    groups.map(({ type, items }) => (
-                        <ProductTypeGroup
-                            key={type.id}
-                            type={type}
+                    groups.map(({ product, items }) => (
+                        <ProductGroup
+                            key={product.id}
+                            product={product}
                             items={items}
                             selected={selected}
                             onToggle={toggle}
-                            onToggleGroup={() => toggleGroupItems(items)}
+                            onToggleGroup={() =>
+                                toggleGroupItems(items.map((item) => item.id))
+                            }
                             onApplyStatus={applyStatus}
                         />
                     ))
                 )}
             </section>
         </AppLayout>
-    );
-}
-
-function ProductTypeGroup({
-    type,
-    items,
-    selected,
-    onToggle,
-    onToggleGroup,
-    onApplyStatus,
-}: {
-    type: ProductTypeWithStatuses;
-    items: TrackingDetail[];
-    selected: number[];
-    onToggle: (id: number) => void;
-    onToggleGroup: () => void;
-    onApplyStatus: (
-        statusId: number,
-        detailIds: number[],
-        statusName?: string,
-    ) => void;
-}) {
-    const [batchStatus, setBatchStatus] = useState<string>('');
-
-    const selectedInGroup = items
-        .map((item) => item.id)
-        .filter((id) => selected.includes(id));
-
-    const allSelected =
-        items.length > 0 && selectedInGroup.length === items.length;
-
-    const nextStatus = (detail: TrackingDetail) =>
-        nextStatusFor(type.statuses, detail.position);
-
-    return (
-        <div className="rounded-xl border border-input">
-            <header className="flex flex-col gap-3 border-b border-input p-4 md:flex-row md:items-center md:justify-between">
-                <div className="flex items-center gap-2">
-                    <h2 className="text-lg font-semibold">
-                        {capitalize(type.name)}
-                    </h2>
-                    <Badge variant="secondary">{items.length}</Badge>
-                    {selectedInGroup.length > 0 && (
-                        <Badge>{selectedInGroup.length} seleccionados</Badge>
-                    )}
-                </div>
-
-                <div className="flex items-center gap-2">
-                    <Select value={batchStatus} onValueChange={setBatchStatus}>
-                        <SelectTrigger className="w-48">
-                            <SelectValue placeholder="Cambiar estado a..." />
-                        </SelectTrigger>
-                        <SelectContent>
-                            {type.statuses.map((status) => (
-                                <SelectItem
-                                    value={String(status.id)}
-                                    key={status.id}
-                                >
-                                    {status.position}. {status.name}
-                                </SelectItem>
-                            ))}
-                        </SelectContent>
-                    </Select>
-                    <Button
-                        disabled={
-                            batchStatus === '' || selectedInGroup.length === 0
-                        }
-                        onClick={() => {
-                            const status = type.statuses.find(
-                                (s) => String(s.id) === batchStatus,
-                            );
-                            onApplyStatus(
-                                Number(batchStatus),
-                                selectedInGroup,
-                                status?.name,
-                            );
-                        }}
-                    >
-                        Aplicar a {selectedInGroup.length}
-                    </Button>
-                </div>
-            </header>
-
-            <Table>
-                <TableHeader>
-                    <TableRow>
-                        <TableHead className="w-[40px]">
-                            <input
-                                type="checkbox"
-                                aria-label="Seleccionar todos"
-                                checked={allSelected}
-                                onChange={onToggleGroup}
-                                className="h-4 w-4 cursor-pointer"
-                            />
-                        </TableHead>
-                        <TableHead>Pedido</TableHead>
-                        <TableHead>Niño / Cliente</TableHead>
-                        <TableHead>Escuela (Aula)</TableHead>
-                        <TableHead>Producto</TableHead>
-                        <TableHead>Notas</TableHead>
-                        <TableHead>Estado</TableHead>
-                        <TableHead>Acciones</TableHead>
-                    </TableRow>
-                </TableHeader>
-                <TableBody>
-                    {items.map((detail) => {
-                        const next = nextStatus(detail);
-
-                        return (
-                            <TableRow
-                                key={detail.id}
-                                className={
-                                    detail.priority
-                                        ? 'bg-amber-50 dark:bg-amber-950/30'
-                                        : undefined
-                                }
-                            >
-                                <TableCell>
-                                    <input
-                                        type="checkbox"
-                                        aria-label={`Seleccionar producto ${detail.product_name}`}
-                                        checked={selected.includes(detail.id)}
-                                        onChange={() => onToggle(detail.id)}
-                                        className="h-4 w-4 cursor-pointer"
-                                    />
-                                </TableCell>
-                                <TableCell>
-                                    <a
-                                        href={route('orders.show', {
-                                            order: detail.order_id,
-                                        })}
-                                        className="underline-offset-2 hover:underline"
-                                    >
-                                        #{detail.order_id}
-                                    </a>
-                                    {detail.photo_number !== null && (
-                                        <span className="ml-1 text-xs text-gray-500">
-                                            (foto {detail.photo_number})
-                                        </span>
-                                    )}
-                                </TableCell>
-                                <TableCell>
-                                    <div className="flex flex-col">
-                                        <span>{detail.child_name ?? '—'}</span>
-                                        <span className="text-xs text-gray-500">
-                                            {detail.client_name}
-                                        </span>
-                                    </div>
-                                </TableCell>
-                                <TableCell>
-                                    {detail.school} (
-                                    {detail.classroom.toUpperCase()})
-                                </TableCell>
-                                <TableCell>
-                                    <div className="flex flex-col">
-                                        <span>{detail.product_name}</span>
-                                        {detail.variant && (
-                                            <span className="text-xs text-gray-500">
-                                                {Object.values(detail.variant)
-                                                    .filter(Boolean)
-                                                    .join(' · ')}
-                                            </span>
-                                        )}
-                                    </div>
-                                </TableCell>
-                                <TableCell className="max-w-[200px] truncate">
-                                    {detail.note}
-                                </TableCell>
-                                <TableCell>
-                                    <div className="flex items-center gap-1">
-                                        {detail.priority && (
-                                            <Badge
-                                                variant="destructive"
-                                                className="gap-1"
-                                            >
-                                                <Flame className="h-3 w-3" />
-                                                Prioridad
-                                            </Badge>
-                                        )}
-                                        <Badge
-                                            variant={
-                                                detail.production_status_id
-                                                    ? 'default'
-                                                    : 'outline'
-                                            }
-                                        >
-                                            {detail.production_status ??
-                                                'Sin empezar'}
-                                        </Badge>
-                                    </div>
-                                    {detail.status_updated_at && (
-                                        <span className="block text-xs text-gray-500">
-                                            {detail.status_updated_at}
-                                        </span>
-                                    )}
-                                </TableCell>
-                                <TableCell>
-                                    {next ? (
-                                        <Button
-                                            size="sm"
-                                            variant="outline"
-                                            title={`Pasar a "${next.name}"`}
-                                            onClick={() =>
-                                                onApplyStatus(
-                                                    next.id,
-                                                    [detail.id],
-                                                    next.name,
-                                                )
-                                            }
-                                        >
-                                            {next.name}
-                                            <ArrowRight className="ml-1 h-4 w-4" />
-                                        </Button>
-                                    ) : (
-                                        <Badge variant="secondary">
-                                            Listo para entregar
-                                        </Badge>
-                                    )}
-                                </TableCell>
-                            </TableRow>
-                        );
-                    })}
-                </TableBody>
-            </Table>
-        </div>
     );
 }
