+++ resources/js/pages/tracking/tests/selection.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { nextStatusFor, toggleGroup, toggleId } from './selection';
+import { nextStatusFor, toggleGroup, toggleId } from '../hooks/use-selection';
 
 describe('toggleId', () => {
     it('adds an unselected id', () => {
